requirements.txt
numpy~=1.25.2
Flask~=2.3.2
Flask-Cors~=4.0.0
ruamel.yaml~=0.17.32
pypresence~=4.3.0
obsws-python~=1.6.0
pandas~=2.0.3
discord-webhook~=1.2.1
Pymem~=1.12.0
psutil~=5.9.5
WMI~=1.5.1
jsonschema~=4.17.3
rich~=13.5.2
cffi~=1.15.1
Pillow~=10.0.1
